refactor: move pure trace parsing types to homeboy-extension-contract
Opens the trace subsystem breakdown (the second-largest extension reverse-edge
cluster, ~45 refs) with the same playbook proven on bench.

Moves the 14 pure trace parsing/result types (TraceStatus, TraceAssertion,
TraceArtifact, TraceCanonicalCheck, TraceEvidenceMetadata, the four *Provenance
types, TraceScenario, TraceList, TraceTemporalAssertionDefinition,
TraceAssertionStatus) plus their pure as_str impls, plus the four span/event
type aliases (TraceEvent/TraceSpanDefinition/TraceSpanResult/TraceSpanStatus =
observation timeline types, already in the leaf lifecycle-contract crate).

TraceResults - the top-level aggregate - stays in core for now; it is blocked
by RigStateSnapshot and TracePreviewMetadata (both reach into crate::rig),
which are the trace equivalent of the sibling blockers cleared during the bench
campaign. Core re-exports the moved types so all trace:: paths stay stable.
